Always apply the default line height in ContextMenu
Normalize the line height in ContextMenu::render itself, next to its
existing rem size and font family normalization, instead of overriding
it at call sites. This fixes the class of bug where menus deferred from
inside the editor inherit a custom buffer_line_height, and removes the
per-site workarounds in layout_mouse_context_menu and the buffer header.

BufferLineHeight moves from theme_settings to the theme crate (with a
re-export) so the ui crate can reference it from the theme layer. The
From<settings::BufferLineHeight> impl becomes a helper function since
it cannot move with the enum.

The completions and code actions popovers keep their override in
element.rs: they are not ui::ContextMenus and size themselves with
eager window.line_height() reads.
