+++ crates/editor/src/element/header.rs
@@ -19,7 +19,6 @@ use smallvec::SmallVec;
 use sum_tree::Bias;
 use text::BufferId;
 use theme::ActiveTheme;
-use theme_settings::BufferLineHeight;
 use ui::{
     ButtonLike, ContextMenu, DiffStat, Indicator, KeyBinding, Tooltip, prelude::*,
     right_click_menu, text_for_keystroke, utils::WithRemSize,
@@ -980,7 +979,7 @@ pub(crate) fn render_buffer_header(
     let editor = editor.clone();
     let buffer_snapshot = buffer.clone();
 
-    let header_with_menu = right_click_menu(("buffer-header-context-menu", buffer_id.to_proto()))
+    right_click_menu(("buffer-header-context-menu", buffer_id.to_proto()))
         .trigger(move |_, _, _| header)
         .menu(move |window, cx| {
             let menu_context = focus_handle.clone();
@@ -1088,16 +1087,7 @@ pub(crate) fn render_buffer_header(
 
                 menu.context(menu_context)
             })
-        });
-
-    // The menu is drawn via a deferred draw, which inherits the editor's text
-    // style, including `buffer_line_height`. Override the line height so the menu
-    // matches context menus elsewhere, as `layout_mouse_context_menu` does for the
-    // editor's mouse context menu.
-    div()
-        .w_full()
-        .line_height(DefiniteLength::Fraction(BufferLineHeight::Comfortable.value()))
-        .child(header_with_menu)
+        })
 }
 
 pub fn file_status_label_color(file_status: Option<FileStatus>) -> Color {
