+++ crates/theme_settings/src/theme_settings.rs
@@ -33,7 +33,8 @@ pub use crate::settings::{
     ThemeAppearanceMode, ThemeName, ThemeSelection, ThemeSettings, adjust_agent_buffer_font_size,
     adjust_agent_ui_font_size, adjust_git_commit_buffer_font_size,
     adjust_markdown_preview_font_size, adjust_ui_font_size, adjusted_font_size, appearance_to_mode,
-    clamp_font_size, default_theme, observe_buffer_font_size_adjustment,
+    buffer_line_height_from_settings, clamp_font_size, default_theme,
+    observe_buffer_font_size_adjustment,
     reset_agent_buffer_font_size, reset_agent_ui_font_size, reset_buffer_font_size,
     reset_git_commit_buffer_font_size, reset_markdown_preview_font_size, reset_ui_font_size,
     set_icon_theme, set_mode, set_theme, setup_ui_font,
